dom/base/test/test_title.html

<!DOCTYPE HTML>
<html>

<head>
  <title>Test for titles</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style type="text/css">
  </style>
</head>

<body onload="runTests()">

<p id="display"></p>
<div style="display:none;">
  <iframe id="html1" src="data:text/html,<html><head><title id='t'>Test</title></head></html>"></iframe>
  <iframe id="html2" src="data:text/html,<html><head><title id='t'>Test</title><title>Foo</title></head></html>"></iframe>
  <iframe id="html3" src="data:text/html,<html></html>"></iframe>
  <iframe id="xhtml1" src="data:text/xml,<html xmlns='http://www.w3.org/1999/xhtml'><body><title id='t'>Test</title></body></html>"></iframe>
  <iframe id="xhtml2" src="data:text/xml,<title xmlns='http://www.w3.org/1999/xhtml'>Test</title>"></iframe>
  <iframe id="xhtml3" src="data:text/xml,<title xmlns='http://www.w3.org/1999/xhtml'>Te<div>bogus</div>st</title>"></iframe>
  <iframe id="xhtml4" src="data:text/xml,<html xmlns='http://www.w3.org/1999/xhtml'/>"></iframe>
  <iframe id="xhtml5" src="data:text/xml,<html xmlns='http://www.w3.org/1999/xhtml'><head/></html>"></iframe>
  <iframe id="xhtml6" src="data:text/xml,<html xmlns='http://www.w3.org/1999/xhtml'><head><style/></head></html>"></iframe>
  <iframe id="xul1" src="data:application/vnd.mozilla.xul+xml,<window xmlns='http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul' title='Test'/>"></iframe>
  <iframe id="xul2" src="data:application/vnd.mozilla.xul+xml,<window xmlns='http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul' title='Test'/>"></iframe>
  <iframe id="svg1" src="data:text/xml,<svg xmlns='http://www.w3.org/2000/svg'><title id='t'>Test</title></svg>"></iframe>
  <iframe id="svg2" src="data:text/xml,<svg xmlns='http://www.w3.org/2000/svg'><title id='t'>Test</title></svg>"></iframe>
</div>

<pre id="test">
<script>
SimpleTest.waitForExplicitFinish();

function runTests() {
  function testStatic(id, expect, description) {
    is(document.getElementById(id).contentDocument.title, expect, description);
  }

  testStatic("html1", "Test", "HTML <title>");
  testStatic("html2", "Test", "choose the first HTML <title>");
  testStatic("html3", "", "No title");
  testStatic("xhtml1", "Test", "XHTML <title> in body");
  testStatic("xhtml2", "Test", "XHTML <title> as root element");
  testStatic("xhtml3", "Test", "XHTML <title> containing an element");
  testStatic("xul1", "Test", "XUL <window> title attribute");
  testStatic("svg1", "Test", "SVG <title>");

  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
